#228 - implement basic plumbing for plugins
This functionality is to be considered in 'alpha' state. The aim is to
enable basic extensibility outside of the Packages/Vintageous directory.

To that end, a PluginManager class has been created to gather date needed
to encode commands as Vintageous expects them.

As a sample implementation of a plugin, check out:
https://github.com/guillermooo/Vintageous_Plugin_Surround