Fix n/N motions after */#
We need to remember the last search command for * and #, and use the command
or the reverse command on n/N.

This requires making _vi_star and _vi_octothorp act in both "grab the current
word and search" and "search for this word" modes.

Also make _vi_star/_vi_octothorp use mode instead of state.mode for testing.