Provide proper support for '^' and '$' anchors
When performing ex-based substitution, '^' and '$' should match the
beginning of a line and the end of a line, respectively. However they
currently match only the beginning and end of the specified range.

Apply the re.MULTILINE flag to the compiled regular expression in order
to enable line-based matching with the '^' and '$' anchors.