Correct range computation for backward movement
The range computed for backward movement in _vi_go_to_line for
INTERNAL_NORMAL mode does not properly account for the case of backward
movement (e.g. when using the sequence d1G). Currently the resulting
range extends from the end of the first line to the beginning of the
last line, when it should be from the beginning of the first to the end
of the last.

Distinguish between the three cases of the start line being the same as
the destination line, the start line coming before the destination line,
and the destination line coming before the start line, and set the range
boundaries accordingly.